Implemented bidding field and button

+++ backend/src/main/java/com/praksa/auction/controller/ProductController.java
@@ -42,7 +42,7 @@ public class ProductController {
     }
 
     @GetMapping("/product/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable("id") long id){
+    public ResponseEntity<Product> getProductById(@PathVariable("id") int id){
         Optional<Product> product = productService.getSelectedProduct(id);
         if(product.isPresent()) return ResponseEntity.ok(product.get());
         else return  new ResponseEntity<>(HttpStatus.NOT_FOUND);
